:truck: Extract customers into a dedicated `customers` app
Move KippoCustomer (model, admin, serializer, viewset, URLs, tests) out of
the `projects` app into a new `customers` app. The DB table is renamed
projects_kippocustomer -> customers_kippocustomer via SeparateDatabaseAndState
so no data is copied and FK constraints follow automatically in PostgreSQL:

- customers/0001_initial: state CreateModel + RunSQL table rename
- projects/0039: state-only AlterField KippoProject.customer ->
  customers.kippocustomer + DeleteModel KippoCustomer

/api/customers/ is unchanged (customers router mounted at /api/ via
SimpleRouter to avoid a duplicate DRF api-root).

The moved KippoCustomerAdmin absorbs two previously-open changes:
- active-project-count column (excludes closed / inactive projects)
- hide the organization field only for single-org users
