+++ kippo/customers/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CustomersConfig(AppConfig):
+    name = "customers"
+    verbose_name = "顧客"
