+++ kippo/projects/admin.py
@@ -14,6 +14,7 @@ from commons.admin import AllowIsStaffAdminMixin, PrettyJSONWidget, UserCreatedB
 from commons.definitions import SATURDAY
 from commons.functions import get_current_month_date_range
 from commons.widgets import MonthYearWidget
+from customers.models import KippoCustomer
 from django import forms
 from django.conf import settings
 from django.contrib import admin, messages
@@ -54,7 +55,6 @@ from .functions import (
 from .models import (
     ActiveKippoProject,
     CollectIssuesAction,
-    KippoCustomer,
     KippoMilestone,
     KippoProject,
     KippoProjectStatus,
@@ -1268,64 +1268,6 @@ class ActiveKippoProjectAdmin(KippoProjectAdmin):
             )
         ).order_by("is_anon_project", "-confidence", "target_date", "name")
         return qs
-
-
-class KippoProjectReadOnlyInline(AllowIsStaffAdminMixin, admin.TabularInline):
-    """Read-only list of projects linked to a KippoCustomer (managed via KippoProjectAdmin)."""
-
-    model = KippoProject
-    fk_name = "customer"
-    extra = 0
-    can_delete = False
-    verbose_name = _("プロジェクト")
-    verbose_name_plural = _("プロジェクト")
-    fields = ("get_project_link", "start_date", "target_date", "billing_date")
-    readonly_fields = ("get_project_link", "start_date", "target_date", "billing_date")
-
-    def has_add_permission(self, request: DjangoRequest, obj: models.Model | None = None) -> bool:  # No Add button
-        return False
-
-    def get_queryset(self, request: DjangoRequest):
-        # earliest target_date first
-        return super().get_queryset(request).order_by("target_date")
-
-    @admin.display(description=KippoProject._meta.get_field("name").verbose_name)
-    def get_project_link(self, obj: KippoProject):
-        return format_html('<a href="{}">{}</a>', obj.get_admin_url(), obj.name)
-
-
-@admin.register(KippoCustomer)
-class KippoCustomerAdmin(AllowIsStaffAdminMixin, UserCreatedBaseModelAdmin):
-    list_display = ("name", "organization", "email", "display_as_active", "updated_datetime")
-    list_display_links = ("name",)
-    list_filter = ("organization", "display_as_active")
-    search_fields = ("name", "email")
-    ordering = ("organization", "-display_as_active", "name")
-    fields = ("organization", "name", "email", "phone", "website", "document_url", "notes", "display_as_active")
-    inlines = (KippoProjectReadOnlyInline,)
-
-    def get_queryset(self, request: DjangoRequest):
-        qs = super().get_queryset(request)
-        if request.user.is_superuser:
-            return qs
-        return qs.filter(organization__in=request.user.organizations).order_by("organization").distinct()
-
-    def get_form(self, request: DjangoRequest, obj: KippoCustomer | None = None, **kwargs) -> Form:
-        form = super().get_form(request, obj, **kwargs)
-        if "organization" not in form.base_fields:
-            return form
-        try:
-            user_initial_organization, _user_organizations = get_user_session_organization(request)
-        except ValueError:
-            user_initial_organization = None
-        if not request.user.is_superuser:
-            form.base_fields["organization"].queryset = request.user.memberships.all()
-        if user_initial_organization:
-            form.base_fields["organization"].initial = user_initial_organization
-            # Hide — derived from the user's session organization. Multi-org users still get
-            # the session value; to create a customer in a different org, switch the session org first.
-            form.base_fields["organization"].widget = forms.HiddenInput()
-        return form
 
 
 @admin.register(KippoMilestone)
