+++ kippo/projects/serializers.py
@@ -14,7 +14,6 @@ from .definitions import (
     ProjectRoles,
 )
 from .models import (
-    KippoCustomer,
     KippoProject,
     KippoProjectUserStatisfactionResult,
     ProjectAssignmentRate,
@@ -25,7 +24,7 @@ from .models import (
 )
 
 if TYPE_CHECKING:
-    from accounts.models import KippoOrganization, OrganizationMembership
+    from accounts.models import OrganizationMembership
 
 
 class ProjectAssignmentRateInlineSerializer(serializers.Serializer):
@@ -148,39 +147,6 @@ class ProjectAssignmentRateSerializer(serializers.ModelSerializer):
         model = ProjectAssignmentRate
         fields = ["id", "project", "project_name", "role", "rate_per_day", "created_datetime", "updated_datetime"]
         read_only_fields = ["id", "project_name", "created_datetime", "updated_datetime"]
-
-
-class KippoCustomerSerializer(serializers.ModelSerializer):
-    """Serializer for KippoCustomer model."""
-
-    organization_name = serializers.CharField(source="organization.name", read_only=True)
-
-    class Meta:
-        model = KippoCustomer
-        fields = [
-            "id",
-            "organization",
-            "organization_name",
-            "name",
-            "email",
-            "phone",
-            "website",
-            "document_url",
-            "notes",
-            "display_as_active",
-            "created_datetime",
-            "updated_datetime",
-        ]
-        read_only_fields = ["id", "organization_name", "created_datetime", "updated_datetime"]
-
-    def validate_organization(self, value: "KippoOrganization") -> "KippoOrganization":
-        request = self.context.get("request")
-        if request is None or request.user.is_superuser:
-            return value
-        user_org_ids = set(request.user.organizationmembership_set.values_list("organization_id", flat=True))
-        if value.id not in user_org_ids:
-            raise serializers.ValidationError("You can only create/update customers in organizations you belong to.")
-        return value
 
 
 class KippoProjectSerializer(serializers.ModelSerializer):
